# docs/openapi/openapi-v2.yaml
openapi: 3.1.0
info:
  title: CircleHaven — Trust Circle API v2 (E2EE X25519/Ed25519)
  version: "2.0.0"
  description: Les corps Auth sont limités à 16 Kio et les corps Messaging à 256 Kio. Les objets d'entrée refusent les propriétés inconnues.
servers:
  - url: https://trust-circle.kavalek.fr
security:
  - bearerAuth: []
    deviceIdHeader: []
    deviceKeyVersionHeader: []
    deviceProofHeader: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token EdDSA/Ed25519 uniquement, audience trust-circle-api et typ=access.
    refreshAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token uniquement, audience trust-circle-auth et typ=refresh.
    deviceIdHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Id
      description: UUID de l'appareil du compte qui produit la preuve d'accès.
    deviceKeyVersionHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Key-Version
      description: Version entière positive de la clé d'identité Ed25519 de l'appareil.
    deviceProofHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Proof
      description: Signature Ed25519 Base64 canonique de la transcription d'accès de 89 octets liée au jti de l'access token. Les trois en-têtes appareil sont obligatoires ensemble sur les routes Messaging, sauf enrôlement initial.
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password, username]
              properties:
                email: { type: string, format: email, minLength: 3, maxLength: 254 }
                password: { type: string, minLength: 8, maxLength: 1024 }
                username: { type: string, minLength: 3, maxLength: 64, pattern: '^(?!\s)(?!.*\s$)[^\u0000-\u001F\u007F]+$' }
      responses:
        "201": { description: Created }
  /auth/login:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email, minLength: 3, maxLength: 254 }
                password: { type: string, minLength: 8, maxLength: 1024 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [access, refresh, user]
                properties:
                  access: { type: string }
                  refresh: { type: string }
                  user: { type: object }
  /auth/refresh:
    post:
      tags: [auth]
      security:
        - refreshAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [access]
                properties:
                  access: { type: string }
  /auth/logout:
    post:
      tags: [auth]
      security:
        - refreshAuth: []
      responses:
        "200": { description: Refresh token révoqué ou session déjà absente }
        "401": { description: Token fourni invalide ou destiné à un autre usage }
  /auth/device-bootstrap-grant:
    post:
      tags: [auth, devices]
      security:
        - bearerAuth: []
      summary: Réautoriser par mot de passe le bootstrap du premier appareil
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password]
              properties:
                password: { type: string, minLength: 8, maxLength: 1024 }
      responses:
        "201":
          description: Grant opaque valable 5 minutes, valeur brute non stockée
          content:
            application/json:
              schema:
                type: object
                required: [grant, expiresAt]
                properties:
                  grant: { type: string, minLength: 43, maxLength: 43, description: Base64URL canonique sans padding de 32 octets CSPRNG }
                  expiresAt: { type: string, format: date-time }
        "401": { description: Access token ou mot de passe invalide }
        "429": { description: Limite de réautorisations atteinte }
...